Create Final_Project
For my final project, I decided to leverage the knowledge I’ve gained in this course to illustrate the correlation between agriculture, supply chains, food security, and life expectancy. The reason I chose this topic is that agriculture has always resonated with me. Growing up on a farm, I developed a deep appreciation for food - seeing the process from planting and harvesting to being served on the table. I’m hopeful that, with advancements in technology, we can help people access food, regardless of where they are, viewing it as a necessity rather than a luxury.

For this assignment, I first looked into the percentage of the population that is undernourished. Then, I explored why certain regions are severely undernourished. Is it because there isn’t enough labor working in the fields? Is it due to a lower allocation of protein production, or is it because insufficient transportation makes it difficult for harvested goods to reach consumers? Finally, I am curious whether there is a direct correlation between food supply and life expectancy.

Disclaimer: 
- The term "Sub-Saharan Africa" used in this project refers to the geographical region of the continent of Africa that lies south of the Sahara.
- All of the data of this assignment is coming from Out World in Data (https://ourworldindata.org/)